chore: make naming consistent
