+++ driver/dm_constant_pad.cpp
@@ -24,7 +24,7 @@
  *
  *******************************************************************************/
 
-#include "constant_pad_driver.hpp"
+#include "pad_constant_driver.hpp"
 #include "registry_driver_maker.hpp"
 
 static Driver* makeDriver(const std::string& base_arg)
